Refactor py2many
