+++ .gitignore
@@ -26,4 +26,6 @@ tests/cases/*.cpp
 tests/cases/*.kt
 tests/cases/*.nim
 dist/
-pyjl/test_files
+tests/output
+target/
+Cargo.lock
