+++ py2many/inference.py
@@ -41,7 +41,7 @@ def infer_types_typpete(node) -> InferMeta:
         infer_types_ast(stmt, context, solver)
 
     solver.push()
-    
+
     return InferMeta(True)
