+++ py2py/transpiler.py
@@ -1,7 +1,8 @@
+from ast import unparse
 import ast
+from py2many.ast_helpers import create_ast_node
 
-from .ast_helpers import create_ast_node, unparse
-from .clike import CLikeTranspiler
+from py2many.clike import CLikeTranspiler
 
 
 class RestoreMainRewriter(ast.NodeTransformer):
