+++ pyjl/transpiler.py
@@ -15,26 +15,9 @@ from .plugins import (
 from py2many.analysis import get_id, is_void_function
 from py2many.declaration_extractor import DeclarationExtractor
 from py2many.clike import _AUTO_INVOKED, class_for_typename
-from py2many.tracer import is_list, defined_before, is_class_or_module, is_enum
+from py2many.tracer import defined_before, is_class_or_module, is_enum
 
 from typing import List, Tuple
-
-
-class JuliaMethodCallRewriter(ast.NodeTransformer):
-    def visit_Call(self, node):
-        fname = node.func
-        if isinstance(fname, ast.Attribute):
-            if is_list(node.func.value) and fname.attr == "append":
-                new_func_name = "push!"
-            else:
-                new_func_name = fname.attr
-            if get_id(fname.value):
-                node0 = ast.Name(id=get_id(fname.value), lineno=node.lineno)
-            else:
-                node0 = fname.value
-            node.args = [node0] + node.args
-            node.func = ast.Name(id=new_func_name, lineno=node.lineno, ctx=fname.ctx)
-        return node
 
 
 class JuliaTranspiler(CLikeTranspiler):
