+++ tests/test_transpile_self.py
@@ -161,23 +161,20 @@ class SelfTranspileTests(unittest.TestCase):
             expected_success={
                 "__main__.py",
                 "analysis.py",
-                "annotation_transformer.py",
                 "ast_helpers.py",
                 "astx.py",
-                "context.py",
                 "cli.py",
+                "context.py",
                 "declaration_extractor.py",
                 "exceptions.py",
                 "language.py",
-                "mutability_transformer.py",
-                "nesting_transformer.py",
-                "python_transformer.py",
                 "result.py",
                 "rewriters.py",
-                "smt.py",
                 "scope.py",
+                "smt.py",
                 "toposort_modules.py",
                 "tracer.py",
+                "transformers.py",
             },
         )
